Push images to GCR instead of GitLab
Since the access token provided by auto-devops is valid only for the
duration of the job, kubernetes will fail to fetch the image if it
tries at any other time (such as when the pod is migrated to another
node or a failed fetch is retried).

The gcloud registry can be configured to allow access with a permanent
service account key, so we'll use that instead.