Created the API end point '/currentuser' which returns a json object representing the current user. This is used in the new underscore templates for the user info at the top of the pages.
Modified log-detail.html to display the Test Owner in the header information if the current user is an admin.
Along with this I removed the test owner information from the 'more' listings.

Fixed adding 'owner' and 'testOwner' to log entries on test creation.

Modified 'index.html' to have 'type="text"' on all the inputs and changed
layout.css so only text fields are indented.