Continue running tests if CheckForScopesInTokenResponse fails
Currently we are seeing this fail, and the conclusion is that it's a
non-compliance in the ping product, so we need to be able to run the
rest of the tests.

It shouldn't have any knock on effects, as far as I can tell we don't
use the returned scopes for anything.