Reduce required entropy
The minimum token length is 128 bits, and expecting to be able to
measure 128 bits of entropy in a 128 bit token is unrealistic.

We typically see values in the 125s. Reduce to 96 for safety.

The entropy is 'shall' so we must only show a failure in the
conformance suite if we're sure the server is not compliant.