Add explicit Accept: application/json headers in various places
This is related to #220, but is not the final position - the main
impetus for this commit is that it may well be needed to get the
server I need to demo on tomorrow to work. The server is believed to
be fussy about the Accept: headers.

This doesn't update /all/ our Accept headers - from a quick check of
logs we still use the default one:

Accept: text/plain, application/json, application/*+json, */*

for at least discovery, and likely other things, but hopefully nothing
important.

I can't currently test that this is sufficient as the server is only
available during working hours Monday-Friday...