Revert "Add userinfo to account request claims (#225)"
This reverts commit 3a5f352e7b18cbb33190239cf7ceb487ec147ce2.

We're still awaiting clarify from OB on the outcome of whether TPPs
are required to make this claim or not, and the current opinion
appears to be that they aren't - so we shouldn't send this claim for
now.

This will mean forgebank will fail, but that's because they're not
compliant with the specs as currently written.