OB: Make some TLS extraction warnings into errors or fatal errors
A user was confused as ExtractMTLSCertificates2FromConfiguration was
failing and only showing as a warning, so they ignored it - but it
meant the test ran on and used the MTLS certificates for client 1
instead, producing confusing results.

A failure here is clearly a hard failure, the test can't run
correctly, so stop.

Change some other warnings to failures, but run on (really as I think
that gives us a chance to still use Dave T's AS for some simple tests).