Test redirect URI with parameters (#202)
Amended from original version Leo did; now when testing the second
client, we automatically use a redirect uri that includes a query
string; this means users need to register two variants of the callback
uri, one of which has ?dummy1=lorem&dummy2=ipsum on the end. The OB
instructions have already been updated to require this.

This firstly checks that redirect uris with query strings are handled
correctly, but also hopefully checks (given that the ? must be percent
encoded in the authorisation request that url encoding of the request
uri is handled correctly).