Fix cases where DisallowTLS10/11 fail unexpectedly
There seems to be a variety of ways in which TLS terminators reject
the less secure TLS methods; cater for a few more we've seen in the
wild.

It still seems like a bad idea to swallow all possible failures as a
"pass" as we can't be sure the test was actually able to run and try
the TLS handshake.

fixes #267