Add trailing whitespace checks to gitlab ci
At the moment we are managing to add/remove trailing whitespace at
sometimes at random, which can result in conflicts / unnecessary
changes showing in diffs at review time / etc. This'll become a become
problem as more developers get involved.

This script will warn about whitespace and is hooked in to the gitlab
ci such that commits that add whitespace will fail the pipeline and
not be mergeable.

checkwhitespace also warns about DOS line endings

This may generate a fake warning if you run this script on windows after
deliberately checking out with DOS line endings, but I'm not sure there's
anything that can realistically be done about that. (It'll be fine when
testing on gitlab as that runs on linux.)

The --fix option can be used to automatically fix errors.