Get selenium to work automatically during local development
Currently the developer has to make some kind of manual tweak to their
system to make sure that selenium works okay when running automated
tests, and the error messages the user gets are not always clear.

Now we get the java app to automatically spawn an instance of redir
that will redirect port 8443 on the java host to the gateway host.

We may technically be breaking some docker rules, however there's
no good easy solutions for getting two processes running in a docker
container without making some noteable changes between the containers
for the local dev system and those on cloud deployments:

https://docs.docker.com/config/containers/multi-service_container/

(The current solution of using extra_hosts in docker-compose.yml to
map localhost on the java server to the httpd can't be automated as
there's no way I can find to get the httpd ip address within
docker-compose.yml.)

This makes no difference at all to the cloud installs.