Allow config to vary for each test in a plan
Configurations can now contain an "override" entry that can override
particular elements the configuration for a particular plan name.

The main purpose of this is to override the browser element so that
we can alter the selenium commands to match what that test will do, e.g.
the negative tests will result in an error page and we can check the
error text is the expected one for that test.

work towards #296