Add IntelliJ project configuration to git
In theory this might make things easier for people who are using IntelliJ.

It's probably not a complete correct configuration as I'm far from an
IntelliJ expert, but it does at least include a debug configuration to
attach to the docker image. I think Steve might've managed to setup
IntelliJ's docker plugin too, but I've not tried that yet.

A number of the xml files are missing newlines at the end, so the
whitespace checker has been updated to ignore the .idea folder.