Add a FAPI-RW test/plan
Conversely to what you might expect this subclasses the OB test and
then overrides a few methods to remove some OB specifics. This is
because making the OB test subclass FAPI is considerably more
difficult due to the hierarchy of OB subclasses we currently have.
Hopefully with time & refactorings this will improve.

This is sufficient to run (and pass) against the FAPI-RW compliant
server Authlete provided.

Note that we are providing an acr claim that is hardcoded; see
https://bitbucket.org/openid/fapi/issues/154/behaviour-of-as-undefined-if-no-acr-claim

If/when the FAPI WG can resolve that discussion we'll need to update
to either not set the acr claim at all, or to allow the user to
configure which one is made.