Fix for OB tests stalling after second browser.goToUrl()
setStatus(WAITING) must always happen on the thread that called
setStatus(RUNNING), as running takes the lock and the lock can only
be released on the thread that claimed it.

Hence we release the lock before scheduling background execution and
claim it again at the start of the background execution.