Clean up setStatus's handling of locks
There were cases where setStatus was being called incorrectly (setting
the status to WAITING on a different thread to the one which set it
to RUNNING). These don't work as we can only unlock the lock on the
thread that locked it; but we were silently ignoring the unlock failing
making such cases difficult to catch.

We tighten up the state machine:

1) Ignore setStatus calls that don't change the status
(I guess ideally these would never happen, but that's a rabiit hole
for another day.)

2) Make sure the expected RUNNING->WAITING transitions
successfully release the lock, by adding a new clearLockIfHeld()
variant that's only used in the unexpected error cases.

3) Throw exceptions on everything except expected transitions

This still leaves remaining problems; at least the test status itself
should be protected by a mutex

Unfortunately these exceptions also don't seem to be logged well in the
test log or the console. Again a problem for another day.