Make browsercontrol able to automatically fill in placeholders
This is a rather evil approach that throws an exception from the
browsercontrol to show that the placeholder should be updated, and
also duplicates some code.

Whilst evil, it works and will allow us to get automated tests in
place - which means we'll then be able to refactor it without being
reliant on manual testing.