Fix docker-compose-dev.yml to work with Safari
For reasons that are unclear to me, the current docker-compose-dev.yml
then prevents me accessing the site in Safari (whereas Chrome is fine).

I think Safari is upset about something in the TLS certificate probably,
although I've gone for what I consider the better solution of making
the -dev httpd configuration identical to the non-dev one, with the
exception of a couple of lines to add the /static/ passthrough.

This seems to make Safari happy.