Fix client secret/code tests having 'PASSED' results
The OpenBanking client secret post/basic and code tests are supposed to
raise warnings; they do correctly put warnings into the logs but if
nothing else goes wrong in the test they leave the status at 'PASSED'.

Update them to call updateResultFromConditionFailure() to fix this,
but that doesn't like being called before the test is created (the
test ends up with result == 'UNKNOWN') so move the calls to the log
functions out of the constructor and into the body of the tests.

Also update the eventLog.log call to use the nicer syntax.