Allow for AS returning an error when we expected an image
It seems there's some ambiguity in the specs as to when the server
should return an error and when it's okay to just display the error to
the user.

This fixes the cases we've seen on the forgerock bank - we now watch
out for an error coming into the redirect uri, and if one arrives we
check it's a valid error response and mark the placeholder as satisfied.

There may be other cases where this happens.

ValidateUserRejectsAuthorizationParametersCorrect probably needs to be
renamed to a general purpose 'verify error response'.

The forgerock bank fails these tests, as it doesn't return the state and
returns an incorrect error.

closes #264