Various updates to .idea
Now uses the latest IntelliJ (2018.2.1) as that seems to solve a few issues
that caused the opening the project from a fresh git checkout to fail in the
older version.

We rename the module in pom.xml as IntelliJ seems to prefer if that matches the
git repository name; not sure if this is really crucial or not.

As per:

https://intellij-support.jetbrains.com/hc/en-us/articles/206544839/comments/360000009220

we enable 'store generated project files externally' as that seems to help it
cleanly open fresh git checkouts.