Treat compiler warnings as error
We've had a few warnings that have been coming out for a while and
there's no reason for it. Update the pom.xml to treat them as errors.

This means adding various @SuppressWarnings("unchecked") annotations
to quieten down the current warnings. We may want to investigate these
sometime, but putting in place a mechanism to stop developers adding
code that generates more warnings is a good start.