Rename state_hash in environment to s_hash
We had a report from a user that it looked like we were checking for
"state_hash" instead of "s_hash". I think it's just confusing to the
user for them to see state_hash in the test results, and I don't see
any strong reason not to use the name used in the id_token as we do
for other hashes.