Fix missing pre-environment & exceptions to cause a failure, never a warning
Currently a missing pre-environment string causes a warning if the
test condition is set to return a warning.

This seems wrong as a required environment value being missing should
be a hard failure - it indicates that the developer who wrote the test
has done something wrong.

If tests want to log a warning when a string is missing from the
environment, they should use skipIfMissing or similar.

Missing pre-requirements or exceptions are also considered a fatal
'test must stop' error, again this seems correct as they indicate
something pretty serious has gone wrong and stopping will make it much
more likely that the developer stops & fixes the problem.