Remove CheckDiscEndpointRequireRequestUriRegistration test
I can't find any convincing argument that this must be set to false
for OB or FAPI profiles.

The original list (
https://gitlab.com/fintechlabs/fapi-conformance-suite/issues/254 )
didn't actually specify that it must be false, just that it was false
in whatever configuration Freddi was looking at.

Potentially we could test that it's at least true or false or absent,
but I'm not sure that's a high priority test right now.