Correct FAPI-R tests to look for error responses in the right place
This also refactors them slightly to use what I think should become
a common pattern - we do nothing in the callback other collect
the fragment, then only resume the test when we have both the query
and the fragment available.

Some of the handling of this callback/fragment collection should probably
be moved into the super class, and I suspect "implicit_hash" should
be renamed to a more generic name; it's just the callback hash that the
conformance suite collected.

work towards #350