Avoid tests that end in warnings showing 1 more warning than happened
We were incorrectly logging a Result value as a 'result' which really
actually expects a ConditionResult value or a TestModule.Status value
of interrupted. We now use TestModule.Status.FINISHED for the result
and add css to display it.

The real result at that time is logged in a new 'testmodule_result' field;
this is manually useful to make it easy to see the original result even
if the user subsequently did something silly that results in unexpected
http arriving at the conformance suite causing the test to move to failed.

Also removes 'interrupted' from the totals at the top of the test page;
as far as I know we can only have exactly one interrupted result, and
if we have one the test status will show as interrupted so there is no
need to have the count.

closes #342