Tweak to allow automated ozone tests to run to completion
The production version of ozone has TLS certificate binding switched off
to make the life of the OB support team easier. This means the certificate
binding test fails and aborts the test, resulting in merge requests
being marked as unmergeable.

As the only problem test is almost the last one in the test, just let
things progress anyway; there doesn't look to be any real assumption
that the certificate binding test has passed in the final few tests.