Allow an empty fragment in ExtractImplicitHashToCallbackResponse
Now that our tests always collect the fragment, it is expected that we
get an empty fragment in the response_type=code tests. Change
ExtractImplicitHash to cope with that case.

This will mean the user won't get an immediate error if the fragment
is empty in a test that should have a fragment, but they'll still
get an error when we start trying to use the response.

(The implicit_hash not being present is now treated, correctly I believe,
as an error by the conformance suite developer.)