Add a custom error page
By the default of Spring Boot, if FAPI has any errors, users will be redirected to the default error page(Whitelable Error Page). That's not good because it makes user feels confuse. Instead of the default error page, user should be redirected to the custom error page. The custom error page shows the error details and helps user easy to back to home page.

Closes #353