First cut at a FAPI-CIBA test using MTLS
Based on the FAPI/OB test that authenticates two clients and checks
that access tokens are correctly bound amongst a bunch of other things.

Currently also includes parts that need to be split out into an OB
profile test.

Targets ping/poll clients.

Testing the ping mode runs into the issues mentioned in #389, in
particular that getting a version of the conformance suite running
locally on my machine to be accessible to the outside world /and/
have an acceptable TLS certificate is... tricky. I've solved this
for now by running an apache reverse proxy in the cloud, which
sends any traffic via https://localhost:3590/ with certificate
checking disabled, and then using ssh -R:3590::8443 <proxy ip> to get
the notification back to my local instance.

There are ~30 outstanding issues that need to be resolved (some
missing tests, some hardcoded strings, some tweaks); they're marked
with FIXMEs and #397 has been opened to make sure those get sorted in
due course.

The automation test check for FAPI-CIBA is disabled as there are a few
issues to solve before we can do automated tests (see #396).