Skip sending of x-fapi-financial-id header by second client in main test
We don't need to add x-fapi-financial-id and x-fapi-auth-date to request headers for second client. Because technically x-fapi-financial-id is no longer mentioned by the standard so does not need to ever be sent, it is fine to leave it being sent. This will also stop any banks rejecting requests that do contain x-fapi-financial-id.

Closes #369