OB+FAPI: Change warnings about non-permitted ciphers to errors
The OB waiver that allows the use of RS256 expires in March, so banks
will have to support PS256 by 13th March according to the latest
updates from OB.

Also update the discovery tests, to require one of PS256 and ES256.
I'm not 100% sure if the UK OB specs allow ES256, but the core FAPI
specifications definitely do.


closes #247