Further tweaks to bad intent id tests
After rereading #274 I realised that the intent of this test was that
the banks server should always redirect back with an error; there is a
correctly signed request object so no reason to doubt the redirect url
is trusted.

The confusion was caused by the fact that the test was creating an
image placeholder - so remove that code.