Create new tests ensure request object without redirect_uri fails
We add the new test ensure request object without "redirect_uri" fails both private_key_jwt and mtls client authentication. We will not send "redirect_uri" in request object for the authorization flow and add "redirect_uri" as query params in the request endpoint url. The expected result is failed for all of these tests.
We will add these tests to test plans (ob-code-id-token-with-mtls and ob-code-id-token-with-private-key-and-matls).
We create new one condition to remove "redirect_uri" from "request_object_claims" which is RemoveRedirectUriFromRequestObject.

Closes #378