Create OB client test that expects a client authentication method to be private_key_jwt.
A new test plan and module created.
`ob-client-test-code-id-token-with-private-key-jwt-and-matls-test-plan`

The task was to validate an encoded client assertion from a client.

Then decode into a readable JWT thus allowing validation of the client authentication method used
relating to the underlying specs for OAuth. (In this case private_key_jwt)
https://tools.ietf.org/html/rfc7523
and
https://openid.net/specs/openid-connect-core-1_0.html#ClientAuthentication

This test currently incorrectly expects the response_type sent by the client to be 'code', when it should be 'code id_token'. That will be addressed in a followup.

This is a clone based off `ob-client-test-code-with-client-secret-basic-and-matls`

MR Relates to #394