Add back the old Openbanking SP tests
Achieved by:

1) checking out commit 99eb84ae841dac563 (which is just before
the removal of the tests)

2) Copy the openbank directory

3) Move back to latest master code

4) Replace openbanking directory with the one from step '2'

5) Use IntelliJ refactor to rename openbanking package to
openbankingdeprecated

6) Tell git to put back the openbanking directory as it is on master

7) Remove OBClientTestCodeWithSecretBasicAndMATLS.java as I don't see a
reason to bring back the client test too

8) remove OBGetResourceEndpoint.java (this is called from test conditions,
so adding an extra one doesn't help us) and revert the places IntelliJ
changed the conditions to call the one in the openbanking deprecated
directory.

This won't build yet, but putting this in a separate commit should make
the rest of the changes easier to review.

part of #402