Add support for v3 OB APIs
This essentially seems to just mean calling a slightly different API and
looking in a slightly different place in the json response when getting
the intent id to go in the request object.

We parse the version number from the API (as the path element is well
defined by the spec) to figure out which version of the API the
configuration is intended to target.

The version number is recorded into the environment so that we look
in the correct place when extracting the answer.

Tested against v3.1 implementation on ozone and seems to work.

part of #402