OB deprecated: Undo the change that skips sending x-fapi-financial-id
The OB v3.1 spec:

https://openbanking.atlassian.net/wiki/spaces/DZ/pages/937656404/Read%2BWrite%2BData%2BAPI%2BSpecification%2B-%2Bv3.1#Read/WriteDataAPISpecification-v3.1-ConsentAuthorisation

still requires that this header is sent. The reason why it does is unclear,
but for safety we should continue sending it in the deprecated tests.

OBSD-6666 has been raised with the OB service desk to query why the
latest OB specs still require it.

related to #369
part of #402