Improve handling of RuntimeExceptions in TestModule
Currently if a runtime exception is thrown in a TestModule, the details
of where the exception happened get lost as we create a new
TestFailureException using only the message from the original exception.

Now when creating the TestFailureException we use the original exception
as the cause, and 'ex()' in DataUtils is updated to include the
stacktrace for the cause in the log entry.

Also tweak some logging.

Spotted after missing out a null check when working on #404