Fix for environments not stopping when branches are deleted
Currently our review apps continue running after the relevant merge
request is merged; I think the intention was always that these should
be tidied up once the code is merged.

According to the gitlab docs:

https://docs.gitlab.com/ee/ci/yaml/#environmentaction

changing the stage to be the same as the stage that created the job
should be enough.

Confusingly, the autodevops template isn't setup like that:

https://gitlab.com/gitlab-org/gitlab-ce/blob/master/lib/gitlab/ci/templates/Auto-DevOps.gitlab-ci.yml

presumably that means it doesn't automatically stop review apps.

Some of the examples do currently follow the documentation though:

https://gitlab.com/gitlab-examples/review-apps-nginx/blob/master/.gitlab-ci.yml