Add a basic integration to pmd, a static analysis tool
The motivation for adding this was to prevent errors creeping into the
code base, e.g. @author tags that we don't want (see #395) and errors
in log format strings that aren't otherwise detected (see
https://gitlab.com/openid/conformance-suite/commit/19293a0e8e617f8c5525564939678196de363f5d#note_266686598
),

Unfortunately the particular error in the logger format string made in
that commit isn't currently detected by PMD:

https://github.com/pmd/pmd/issues/2196

We add:
- a custom ruleset, .pmd.releset.xml
- mvn configuration to check this ruleset when running tests
- intellij pdm plugin config to run this ruleset
- Add custom pmd rule to flag up use of @author (part of #395)

The mvn configuration uses the most recent version of pmd currently
available (6.20.0).

We disable the PMD CPD (copy & paste detector) for now, as it produces
a lot on errors on our current codebase.

There's probably a lot more we could do, but this is a start.