FAPI-OB: Change discovery hostname check to a warning
Currently a url that is not on the OB production directory
causes a test failure, change that down to a warning as we
want non-production systems to be certifiable. As we still
want other JWKS validation failures to be an error this means
splitting the hostname validation out into a separate
condition. As the hostname validation is only used for this
check, move it down into the concrete class so we can make
the error message much more specific.

Also refactor/cleanup ValidateJsonUri a bit, including making
it a proper abstract class.