Minor wording change for unnecessary image placeholders
Sometimes we don't know where a test will result in a redirect back
to the conformance suite or the user uploading an image - in these
cases we create a placeholder, then after we get the redirect we
mark the placeholder as no longer needed. Just change the wording
used there to make it hopefully more obvious to someone reviewing
the log.

This confused OIDF when reviewing logs submitted for certification.