Move test to 'REVIEW' status if an image placeholder was filled
Reviewing certification logs with the OIDF team made it clear that it
is desireable to know when a test fully passed, and when it 'passed'
due to a browsercontrol automation filling in a placeholder (which
from a certification point of view may still require the team to
manually check the returned error looks correct).

As browser control doesn't have direct access to the testRunner/testModule,
we solve this by doing a check of the placeholders when the test finishes,
just before the existing code that marks them as satisified if the test
completes.

closes #434