Negative Test : Add expired iat value ID token Add expired iat value to id_token. Thus create a fail case to check that the client receives an invalid id_token from the authorization endpoint, and hence shouldn't call the token endpoint.
The client will error;

For expired iat: "iat is too far in the past".

Requires sample-openbanking-client-nodejs!12

The conformance-suite should also be able to handle a client that does not verify the status of the iat value, hence callAndStopOnFailure(AddInvalidExpiredIatValueToIdToken.class, ConditionResult.WARNING). Alerting the user that an invalid iat value was created in the id_token and the client allows for it to continue with the authorization flow.

https://openid.net/specs/openid-connect-core-1_0.html#IDTokenValidation - Section 3.1.3.7-10

Relates to #419

Additionally, create missing FAPI-RW exp test modules:

*  "fapi-rw-id2-client-test-with-mtls-holder-of-key-invalid-missing-exp"
*  "fapi-rw-id2-client-test-with-mtls-holder-of-key-invalid-expired-exp"