Positive Test : Add aud value as array to ID token
Add the aud value as an array instead of a string to the id_token.
Thus create an OpenID connect flow case to validate that having the aud value as an array of length 1, is permitted.

Refactor SignIdToken so that we can perform multiple methods of signing the id_token,
hence the creation of SignIdTokenBypassingNimbusChecks.

The purpose is to bypass the traditional signing method of the JWTClaimSet that prevents a user from creating a single entry aud array.
https://gitlab.com/openid/conformance-suite/merge_requests/486/diffs#00e608067fd96465d2f03565779c5458fcf45d45_115_113

Instead, we utilize the superclass of SignedJWT - JWSObject.
Populate the header and payload, and sign with JWSSigner.

Relates to #419