FAPI Client Test: Skip if state is not supplied by client
Allow the server to handle a client NOT sending a state value.
This is because this is OPTIONAL as per the spec -  https://openid.net/specs/openid-financial-api-part-2-ID2.html.

Additionally, the server now exits early for the tests concerning 'missing s_hash' - as it cannot test if state is missing.

Closes #442